requirements.txt
arrow==1.0.0
astroid==2.4.2
atomicwrites==1.4.0
attrs==20.3.0
autopep8==1.5.7
Brotli==1.0.9
cffi==1.14.5
click==7.1.2
colorama==0.4.4
cryptography==3.4.7
Flask==1.1.2
Flask-Compress==1.8.0
Flask-JWT==0.3.2
Flask-WTF==0.14.3
future==0.18.2
gunicorn==20.0.4
idna==2.10
iniconfig==1.1.1
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
names==0.3.0
packaging==20.9
pathlib==1.0.1
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
PyJWT==1.4.2
pylint==2.6.0
pyparsing==2.4.7
pytest==6.2.3
python-dateutil==2.8.1
random-string==1.0
random-text-generator==1.1.0
six==1.15.0
toml==0.10.2
url64==0.0.2
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
